fixes bugs centered around creating new projects and renaming them.
A bug was deleting all projects because the input event listener was clearing by an index of -1 when recieving the incorrect input, therefore deleting the wrong data.

Fixed by adding a check to make sure the index exists.

A bug still exists that will delete any saved list if it's name is typed into the input field.